requirements.txt
streamlit>=1.24.0
tensorflow>=2.8.0
keras-tcn>=3.4.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.13.0
pyserial>=3.5 